+++ packages/ordia-core/ordia/__init__.py
@@ -8,9 +8,9 @@ try:
     try:
         __version__ = version("ordia-core")
     except PackageNotFoundError:
-        __version__ = "0.15.1"
+        __version__ = "0.18.0"
 except ImportError:  # pragma: no cover
-    __version__ = "0.15.1"
+    __version__ = "0.18.0"
 
 __all__ = [
     "OrdiaConfig",